# frontend/public/openapi.yaml
openapi: 3.1.0
info:
  title: GenQMA Intelligence & Payments API
  description: '# GenQMA Intelligence & Payments API


    GenQMA is a provider marketplace for quantitative market intelligence. Clients
    discover live signals, select a provider-bound report, create a query-bound invoice,
    verify Circle Gateway x402 settlement, and retrieve the resulting preview or full
    report.


    ## Recommended buyer flow

    1. Discover providers with `GET /api/v1/providers` or live candidates with `GET
    /api/v1/agent/recommendations`.

    2. Optional: call `POST /api/v1/payment/quote` for a provider-bound price.

    3. Create an invoice with `POST /api/v1/payment/invoice`.

    4. Pay the returned Circle Gateway requirement, then verify with `POST /api/v1/payment/verify`.

    5. Poll `GET /api/v1/payment/invoices/{invoice_id}/status` when settlement is
    still processing.

    6. Retrieve the paid provider report from the provider-specific Reports route.


    ## Agent flow

    Use `POST /api/v1/agent/decision` to turn a bounded budget and provider/tier policy
    into a validated purchase decision. The payment executor remains responsible for
    signing and submitting payment; this API does not custody browser wallet keys.


    ## Access levels

    - **Public:** no credential is required. Optional-auth operations return additional
    private data only after verification.

    - **Protected:** requires a paid-access token, wallet-owner token, invoice secret,
...

...

...


...

...
servers:
...
paths:
...
  /api/v1/providers/{provider_id}/preview:
    post:
      tags:
      - Reports
      summary: Provider Preview Signal
      description: Returns a paid provider preview for the exact query snapshot bound
        to the invoice.
      operationId: provider_preview_signal_api_v1_providers__provider_id__preview_post
      security:
      - X-QMA-Access-Token: []
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          description: Quantitative intelligence provider identifier (e.g. 'funding_memory',
            'polymarket_orderbook', 'pyth_entropy').
          examples:
          - funding_memory
          title: Provider Id
        description: Quantitative intelligence provider identifier (e.g. 'funding_memory',
          'polymarket_orderbook', 'pyth_entropy').
      - name: invoice_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional settled invoice ID for pre-allocated invoice flows.
            Leave empty when paying directly via x402 / Circle Gateway.
          examples:
          - inv_79d896a28cd5
          title: Invoice Id
        description: Optional settled invoice ID for pre-allocated invoice flows.
          Leave empty when paying directly via x402 / Circle Gateway.
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/QueryModel'
              - type: 'null'
              description: Quantitative analysis query parameters including symbol
                and market metrics.
              title: Query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderReportResponse'
        '400':
          description: The request is invalid or violates a business rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Payment is required or the invoice is not settled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The caller is not authorized for this resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: The request was rate limited. Retry after the indicated delay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitErrorResponse'
        '500':
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: paid-access
      x-qma-audiences:
      - agent
      - private
      - wallet
      x-payment-info: *id001
  /api/v1/providers/{provider_id}/full-report:
    post:
      tags:
      - Reports
      summary: Provider Full Report
      description: Returns a paid provider full report for the exact query snapshot
        bound to the invoice.
      operationId: provider_full_report_api_v1_providers__provider_id__full_report_post
      security:
      - X-QMA-Access-Token: []
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          description: Quantitative intelligence provider identifier (e.g. 'funding_memory',
            'polymarket_orderbook', 'pyth_entropy').
          examples:
          - funding_memory
          title: Provider Id
        description: Quantitative intelligence provider identifier (e.g. 'funding_memory',
          'polymarket_orderbook', 'pyth_entropy').
      - name: invoice_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional settled invoice ID for pre-allocated invoice flows.
            Leave empty when paying directly via x402 / Circle Gateway.
          examples:
          - inv_79d896a28cd5
          title: Invoice Id
        description: Optional settled invoice ID for pre-allocated invoice flows.
          Leave empty when paying directly via x402 / Circle Gateway.
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/QueryModel'
              - type: 'null'
              description: Quantitative analysis query parameters including symbol
                and market metrics.
              title: Query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderReportResponse'
        '400':
          description: The request is invalid or violates a business rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Payment is required or the invoice is not settled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The caller is not authorized for this resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: The request was rate limited. Retry after the indicated delay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitErrorResponse'
        '500':
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: paid-access
      x-qma-audiences:
      - agent
      - private
      - wallet
      x-payment-info: *id001
  /api/v1/preview:
    post:
      tags:
      - Legacy compatibility
      summary: Preview Signal
      description: Backward-compatible Funding Memory preview endpoint.
      operationId: preview_signal_api_v1_preview_post
      deprecated: true
      security:
      - X-QMA-Access-Token: []
      parameters:
      - name: invoice_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional settled invoice ID for pre-allocated invoice flows.
            Leave empty when paying directly via x402 / Circle Gateway.
          title: Invoice Id
        description: Optional settled invoice ID for pre-allocated invoice flows.
          Leave empty when paying directly via x402 / Circle Gateway.
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/QueryModel'
              - type: 'null'
              title: Query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderReportResponse'
        '400':
          description: The request is invalid or violates a business rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Payment is required or the invoice is not settled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The caller is not authorized for this resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: The request was rate limited. Retry after the indicated delay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitErrorResponse'
        '500':
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: paid-access
      x-qma-audiences:
      - private
      x-payment-info: *id001
  /api/v1/analyze:
    post:
      tags:
      - Legacy compatibility
      summary: Analyze Signal
      description: Backward-compatible Funding Memory full report endpoint.
      operationId: analyze_signal_api_v1_analyze_post
      deprecated: true
      security:
      - X-QMA-Access-Token: []
      parameters:
      - name: invoice_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional settled invoice ID for pre-allocated invoice flows.
            Leave empty when paying directly via x402 / Circle Gateway.
          title: Invoice Id
        description: Optional settled invoice ID for pre-allocated invoice flows.
          Leave empty when paying directly via x402 / Circle Gateway.
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/QueryModel'
              - type: 'null'
              title: Query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderReportResponse'
        '400':
          description: The request is invalid or violates a business rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Payment is required or the invoice is not settled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The caller is not authorized for this resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: The request was rate limited. Retry after the indicated delay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitErrorResponse'
        '500':
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: paid-access
      x-qma-audiences:
      - private
      x-payment-info: *id001
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
        message:
          type: string
          title: Message
        status_code:
          type: integer
          title: Status Code
        detail:
          anyOf:
          - {}
          - type: 'null'
          title: Detail
      additionalProperties: true
      type: object
      required:
      - error
      - message
      - status_code
      title: ErrorResponse
      description: 'Error envelope returned for handled HTTP failures.


        ``detail`` is retained for compatibility with existing QMA clients while

        ``error`` and ``message`` provide stable fields for new clients and agents.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProviderReportResponse:
      properties:
        query_symbol:
          anyOf:
          - type: string
          - type: 'null'
          title: Query Symbol
        query:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Query
        query_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Query Hash
        tier:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier
        upgrade_cta:
          anyOf:
          - type: string
          - type: 'null'
          title: Upgrade Cta
        invoice:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Invoice
        provider_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Note
        provider_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Id
        provider_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Name
        provider_owner_wallet:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Owner Wallet
        paid_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Paid At
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
        funding_context:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Funding Context
        regime_cluster:
          anyOf:
          - type: string
          - type: 'null'
          title: Regime Cluster
        regime_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Regime Description
        is_ood:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Ood
        ood_p_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Ood P Value
        win_rate_band:
          anyOf:
          - type: string
          - type: 'null'
          title: Win Rate Band
        rough_win_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Rough Win Rate
        top_analogs:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Top Analogs
        analysis_focus:
          anyOf:
          - type: string
          - type: 'null'
          title: Analysis Focus
        turnover_context:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Turnover Context
        provider_diagnostics:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Provider Diagnostics
        provider_specific_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Provider Specific Data
          description: Provider-specific report data (Legacy).
      additionalProperties: true
      type: object
      title: ProviderReportResponse
      description: Paid report envelope with extensible provider-specific report fields.
...
    QueryModel:
      properties:
        symbol:
          type: string
          maxLength: 32
          minLength: 1
          title: Symbol
          description: Cryptocurrency or perpetual market asset ticker pair (e.g.
            'BTC_USDT', 'ETH_USDT').
          examples:
          - BTC_USDT
        fundingRate:
          anyOf:
          - type: number
          - type: 'null'
          title: Fundingrate
          description: Current 8-hour perpetual contract funding rate as a decimal
            float (e.g. -0.00042 for negative rate).
          default: 0.0
          examples:
          - -0.00042
        marketCap:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Marketcap
          description: Total circulating market capitalization denominated in USD.
          examples:
          - 1250000000
        FDV:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Fdv
          description: Fully diluted valuation (FDV) denominated in USD.
          examples:
          - 1800000000
        circRatio:
          anyOf:
          - type: number
            maximum: 1.5
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Circratio
          description: Ratio of circulating supply to maximum/total supply (between
            0.0 and 1.5).
          examples:
          - 0.69
        fromATH:
          anyOf:
          - type: number
          - type: 'null'
          title: Fromath
          description: Drawdown percentage from the asset's historical all-time high
            (negative value, e.g. -42.5).
          examples:
          - -42.5
        volume24h:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Volume24H
          description: Aggregated 24-hour global trading volume denominated in USD.
          examples:
          - 85400000
        amount:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Amount
          description: Simulated order book trade size or liquidity depth test in
            USD.
          examples:
          - 125000
        openInterest:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Openinterest
          description: Aggregate open interest in perpetual contracts denominated
            in USD.
          examples:
          - 62500000
        openInterestChange24h:
          anyOf:
          - type: number
          - type: 'null'
          title: Openinterestchange24H
          description: 24-hour percentage change in open interest (e.g. 8.4 for +8.4%).
          examples:
          - 8.4
        longShortRatio:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Longshortratio
          description: Ratio of top trader long accounts versus short accounts (e.g.
            1.12).
          examples:
          - 1.12
        price:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Price
          description: Current spot index or perpetual mark price in USD.
          examples:
          - 67250.5
      additionalProperties: true
      type: object
      required:
      - symbol
      title: QueryModel
...
    RateLimitErrorResponse:
      properties:
        error:
          type: string
          title: Error
        message:
          type: string
          title: Message
        status_code:
          type: integer
          title: Status Code
        detail:
          anyOf:
          - {}
          - type: 'null'
          title: Detail
        scope:
          type: string
          title: Scope
        limit:
          type: integer
          title: Limit
        window_seconds:
          type: integer
          title: Window Seconds
        retry_after_seconds:
          anyOf:
          - type: integer
          - type: number
          title: Retry After Seconds
      additionalProperties: true
      type: object
      required:
      - error
      - message
      - status_code
      - scope
      - limit
      - window_seconds
      - retry_after_seconds
      title: RateLimitErrorResponse
      description: Error envelope returned when the API rate limiter rejects a request.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    x-qma-admin-token:
      type: apiKey
      description: 'Static administrative token for platform operations.


        **Enforcement:**

        - **Required:** `POST /api/v1/providers/{provider_id}/toggle` and `POST /api/v1/creators/applications/{application_id}/review`.

        - **Required for an unfiltered application list:** `GET /api/v1/creators/applications`
        accepts an admin token to list all applications; without it, callers must
        provide a wallet filter.

        - **Conditional:** provider discovery/stats require the token only when `include_disabled=true`.

        Invalid or missing credentials on a protected admin operation return `403
        Forbidden`.'
      in: header
      name: x-qma-admin-token
    X-QMA-Access-Token:
      type: apiKey
      description: 'Short-lived JWT granted after a successful invoice settlement.
        Used to unlock paid report endpoints.


        **Retrieval:** A token can be returned by payment verification/status after
        the invoice reaches the server-accepted paid state.

        **Enforcement:** Required on paid report endpoints and `POST /api/v1/chat`
        via the `X-QMA-Access-Token` header. Missing, invalid, expired, or invoice-mismatched
        tokens trigger `403 Forbidden`.'
      in: header
      name: X-QMA-Access-Token
    X-QMA-Wallet-Token:
      type: apiKey
      description: 'Wallet profile session JWT, generated by signing a SIWE-like message
        proving wallet ownership.


        **Retrieval:** Returned by the `/api/v1/wallets/{address}/session` endpoint.
...